refactor(region map): extract sampling helpers to region_map_sampling.js and update imports
